passwordstrength.php
passwordstrength.php is a function I wrote to work in conjunction with my generate password function "genpass.php". It will rate a password based on some criteria and return a percentage rating, error messages, and an "assessment" message (like "great!" or "awful!"). I'm adding it to the repository as it is now in its current state.