testgenpass.php
testgenpass.php contains a handy HTML form that allows the user to test the functionality of the two functions "genpass.php" and "passwordstrength.php"

This page allows the user to quickly generate passwords that meet any set of specified criteria, and test the password strength at the same time. The user can specify the overall length of the password, as well as how many numbers, lower case letters, upper case letters, and special characters should or should not be included in the password. There is also an option to force all of the characters in the password to be unique.